+++ sdk/cosmos/azure-cosmos/test/test_service_retry_policies_async.py
@@ -11,7 +11,7 @@ from azure.core.exceptions import ServiceRequestError, ServiceResponseError
 
 import test_config
 from azure.cosmos import DatabaseAccount, _location_cache
-from azure.cosmos._location_cache import RegionalEndpoint
+from azure.cosmos._location_cache import DualEndpoint
 from azure.cosmos.aio import CosmosClient, _retry_utility_async, _global_endpoint_manager_async
 from azure.cosmos.exceptions import CosmosHttpResponseError
 
@@ -26,7 +26,7 @@ class TestServiceRetryPoliciesAsync(unittest.IsolatedAsyncioTestCase):
     REGION1 = "West US"
     REGION2 = "East US"
     REGION3 = "West US 2"
-    REGIONAL_ENDPOINT = RegionalEndpoint(host, host)
+    REGIONAL_ENDPOINT = DualEndpoint(host, host)
 
     @classmethod
     def setUpClass(cls):
@@ -54,13 +54,13 @@ class TestServiceRetryPoliciesAsync(unittest.IsolatedAsyncioTestCase):
 
             # Change the location cache to have 3 preferred read regions and 3 available read endpoints by location
             original_location_cache = mock_client.client_connection._global_endpoint_manager.location_cache
-            original_location_cache.available_read_locations = [self.REGION1, self.REGION2, self.REGION3]
+            original_location_cache.account_read_locations = [self.REGION1, self.REGION2, self.REGION3]
             original_location_cache.available_read_regional_endpoints_by_locations = {
                 self.REGION1: self.REGIONAL_ENDPOINT,
                 self.REGION2: self.REGIONAL_ENDPOINT,
                 self.REGION3: self.REGIONAL_ENDPOINT}
-            original_location_cache.read_regional_endpoints = [self.REGIONAL_ENDPOINT, self.REGIONAL_ENDPOINT,
-                                                               self.REGIONAL_ENDPOINT]
+            original_location_cache.read_dual_endpoints = [self.REGIONAL_ENDPOINT, self.REGIONAL_ENDPOINT,
+                                                           self.REGIONAL_ENDPOINT]
             try:
                 # Mock the function to return the ServiceRequestException we retry
                 mf = self.MockExecuteServiceRequestException()
@@ -73,8 +73,8 @@ class TestServiceRetryPoliciesAsync(unittest.IsolatedAsyncioTestCase):
                 _retry_utility_async.ExecuteFunctionAsync = self.original_execute_function
 
             # Now we change the location cache to have only 1 preferred read region
-            original_location_cache.available_read_locations = [self.REGION1]
-            original_location_cache.read_regional_endpoints = [self.REGIONAL_ENDPOINT]
+            original_location_cache.account_read_locations = [self.REGION1]
+            original_location_cache.read_dual_endpoints = [self.REGIONAL_ENDPOINT]
             try:
                 # Reset the function to reset the counter
                 mf = self.MockExecuteServiceRequestException()
@@ -87,8 +87,8 @@ class TestServiceRetryPoliciesAsync(unittest.IsolatedAsyncioTestCase):
                 _retry_utility_async.ExecuteFunctionAsync = self.original_execute_function
 
             # Now we try it out with a write request
-            original_location_cache.available_write_locations = [self.REGION1, self.REGION2]
-            original_location_cache.write_regional_endpoints = [self.REGIONAL_ENDPOINT, self.REGIONAL_ENDPOINT]
+            original_location_cache.account_write_locations = [self.REGION1, self.REGION2]
+            original_location_cache.write_dual_endpoints = [self.REGIONAL_ENDPOINT, self.REGIONAL_ENDPOINT]
             original_location_cache.available_write_regional_endpoints_by_locations = {
                 self.REGION1: self.REGIONAL_ENDPOINT,
                 self.REGION2: self.REGIONAL_ENDPOINT}
@@ -116,13 +116,13 @@ class TestServiceRetryPoliciesAsync(unittest.IsolatedAsyncioTestCase):
 
             # Change the location cache to have 3 preferred read regions and 3 available read endpoints by location
             original_location_cache = mock_client.client_connection._global_endpoint_manager.location_cache
-            original_location_cache.available_read_locations = [self.REGION1, self.REGION2, self.REGION3]
+            original_location_cache.account_read_locations = [self.REGION1, self.REGION2, self.REGION3]
             original_location_cache.available_read_regional_endpoints_by_locations = {
                 self.REGION1: self.REGIONAL_ENDPOINT,
                 self.REGION2: self.REGIONAL_ENDPOINT,
                 self.REGION3: self.REGIONAL_ENDPOINT}
-            original_location_cache.read_regional_endpoints = [self.REGIONAL_ENDPOINT, self.REGIONAL_ENDPOINT,
-                                                               self.REGIONAL_ENDPOINT]
+            original_location_cache.read_dual_endpoints = [self.REGIONAL_ENDPOINT, self.REGIONAL_ENDPOINT,
+                                                           self.REGIONAL_ENDPOINT]
             try:
                 # Mock the function to return the ClientConnectionError we retry
                 mf = self.MockExecuteServiceResponseException(AttributeError, None)
@@ -135,8 +135,8 @@ class TestServiceRetryPoliciesAsync(unittest.IsolatedAsyncioTestCase):
                 _retry_utility_async.ExecuteFunctionAsync = self.original_execute_function
 
             # Now we change the location cache to have only 1 preferred read region
-            original_location_cache.available_read_locations = [self.REGION1]
-            original_location_cache.read_regional_endpoints = [self.REGIONAL_ENDPOINT]
+            original_location_cache.account_read_locations = [self.REGION1]
+            original_location_cache.read_dual_endpoints = [self.REGIONAL_ENDPOINT]
             try:
                 # Reset the function to reset the counter
                 mf = self.MockExecuteServiceResponseException(AttributeError, None)
@@ -149,8 +149,8 @@ class TestServiceRetryPoliciesAsync(unittest.IsolatedAsyncioTestCase):
                 _retry_utility_async.ExecuteFunctionAsync = self.original_execute_function
 
             # Now we try it out with a write request
-            original_location_cache.available_write_locations = [self.REGION1, self.REGION2]
-            original_location_cache.write_regional_endpoints = [self.REGIONAL_ENDPOINT, self.REGIONAL_ENDPOINT]
+            original_location_cache.account_write_locations = [self.REGION1, self.REGION2]
+            original_location_cache.write_dual_endpoints = [self.REGIONAL_ENDPOINT, self.REGIONAL_ENDPOINT]
             original_location_cache.available_write_regional_endpoints_by_locations = {
                 self.REGION1: self.REGIONAL_ENDPOINT,
                 self.REGION2: self.REGIONAL_ENDPOINT}
@@ -169,8 +169,8 @@ class TestServiceRetryPoliciesAsync(unittest.IsolatedAsyncioTestCase):
 
             # If we do a write request with a ClientConnectionError,
             # we will do cross-region retries like with read requests
-            original_location_cache.available_write_locations = [self.REGION1, self.REGION2]
-            original_location_cache.write_regional_endpoints = [self.REGIONAL_ENDPOINT, self.REGIONAL_ENDPOINT]
+            original_location_cache.account_write_locations = [self.REGION1, self.REGION2]
+            original_location_cache.write_dual_endpoints = [self.REGIONAL_ENDPOINT, self.REGIONAL_ENDPOINT]
             original_location_cache.available_write_regional_endpoints_by_locations = {
                 self.REGION1: self.REGIONAL_ENDPOINT,
                 self.REGION2: self.REGIONAL_ENDPOINT}
@@ -246,15 +246,15 @@ class TestServiceRetryPoliciesAsync(unittest.IsolatedAsyncioTestCase):
             await container.read()
 
             original_location_cache = mock_client.client_connection._global_endpoint_manager.location_cache
-            original_location_cache.available_read_locations = [self.REGION1, self.REGION2, self.REGION3]
+            original_location_cache.account_read_locations = [self.REGION1, self.REGION2, self.REGION3]
             original_location_cache.available_read_regional_endpoints_by_locations = {
                 self.REGION1: self.REGIONAL_ENDPOINT,
                 self.REGION2: self.REGIONAL_ENDPOINT,
                 self.REGION3: self.REGIONAL_ENDPOINT}
-            original_location_cache.read_regional_endpoints = [self.REGIONAL_ENDPOINT, self.REGIONAL_ENDPOINT,
-                                                               self.REGIONAL_ENDPOINT]
-            original_location_cache.available_write_locations = [self.REGION1, self.REGION2]
-            original_location_cache.write_regional_endpoints = [self.REGIONAL_ENDPOINT, self.REGIONAL_ENDPOINT]
+            original_location_cache.read_dual_endpoints = [self.REGIONAL_ENDPOINT, self.REGIONAL_ENDPOINT,
+                                                           self.REGIONAL_ENDPOINT]
+            original_location_cache.account_write_locations = [self.REGION1, self.REGION2]
+            original_location_cache.write_dual_endpoints = [self.REGIONAL_ENDPOINT, self.REGIONAL_ENDPOINT]
             original_location_cache.available_write_regional_endpoints_by_locations = {
                 self.REGION1: self.REGIONAL_ENDPOINT,
                 self.REGION2: self.REGIONAL_ENDPOINT}
@@ -307,7 +307,7 @@ class TestServiceRetryPoliciesAsync(unittest.IsolatedAsyncioTestCase):
 
             # Reset the location cache's unavailable endpoints in order to try the same with other exceptions
             original_location_cache.location_unavailability_info_by_endpoint = {}
-            original_location_cache.write_regional_endpoints = [self.REGIONAL_ENDPOINT, self.REGIONAL_ENDPOINT]
+            original_location_cache.write_dual_endpoints = [self.REGIONAL_ENDPOINT, self.REGIONAL_ENDPOINT]
 
             # Now we test ClientConnectionResetError, the subclass of ClientConnectionError
             try:
@@ -328,7 +328,7 @@ class TestServiceRetryPoliciesAsync(unittest.IsolatedAsyncioTestCase):
 
             # Reset the location cache's unavailable endpoints in order to try the same with other exceptions
             original_location_cache.location_unavailability_info_by_endpoint = {}
-            original_location_cache.write_regional_endpoints = [self.REGIONAL_ENDPOINT, self.REGIONAL_ENDPOINT]
+            original_location_cache.write_dual_endpoints = [self.REGIONAL_ENDPOINT, self.REGIONAL_ENDPOINT]
 
             # Now we test ServerConnectionError, the subclass of ClientConnectionError
             try:
@@ -344,7 +344,7 @@ class TestServiceRetryPoliciesAsync(unittest.IsolatedAsyncioTestCase):
 
             # Reset the location cache's unavailable endpoints in order to try the same with other exceptions
             original_location_cache.location_unavailability_info_by_endpoint = {}
-            original_location_cache.write_regional_endpoints = [self.REGIONAL_ENDPOINT, self.REGIONAL_ENDPOINT]
+            original_location_cache.write_dual_endpoints = [self.REGIONAL_ENDPOINT, self.REGIONAL_ENDPOINT]
 
             # Now we test ClientOSError, the subclass of ClientConnectionError
             try: